services/oanda_stream/requirements.txt
google-cloud-error-reporting==1.0.*
requests==2.24.*
pymongo==3.11.*
requests==2.27.*